Allow pagination with `meta` key in response
The `live_assessment` endpoint in Canvas returns pagination values
in the response body rather than in a `<Link>` header. See #605
for more detail.

This update checks first for a `Link` header and then for a `meta`
property in the response when setting the `next_link` value while
iterating the response from Canvas.

This method relies on the `_root` property being set by the user to
specify which key holds the target data from Canvas.